Fixed view state
Previously, there was no function to clear the view and push a base, resorting in accidental memory leaks. This has been fixed with a new `enginesetview()` function. Main menu and Sandbox views can now be swapped between. Additionally, fixed engine's lack of clearing and refreshing the terminal. This could was in the Sandbox view, which should not at all determine this behaviour. Engine functions have been renamed to feel more natural when typing.